rearrange state, prettify prototype
